[#164] Refactor: 홈 공용 쿼리 훅을  fridge/food 도메인으로 분리
