+++ src/features/food/hooks/queries/useFoodStatusQuery.ts
@@ -1,6 +1,6 @@
+import { getFoodStatusApi } from "@/features/food/apis/foodStatus";
 import { useApiQuery } from "@/shared/apis/builder/ApiBuilder";
 import { QUERY_KEYS } from "@/shared/constants/queryKeys";
-import { getFoodStatusApi } from "../../apis/food";
 
 const useFoodStatusQuery = (fridgeId: number) => {
   return useApiQuery(getFoodStatusApi, QUERY_KEYS.food.status(fridgeId), {
