+++ src/features/fridge/__tests__/hooks/useFridgeQuery.test.tsx
@@ -4,6 +4,10 @@ import { renderHook, waitFor } from "@testing-library/react-native";
 import React from "react";
 import { useFridgeQuery } from "../../hooks/queries/useFridgeQuery";
 
+jest.mock("@/features/auth/store/useAuthStore", () => ({
+  useIsLoggedIn: () => true,
+}));
+
 jest.mock("@/shared/apis/apiClient");
 const mockedApiClient = apiClient as jest.MockedFunction<typeof apiClient>;
