+++ src/features/fridge/hooks/queries/useFridgeQuery.ts
@@ -1,5 +1,5 @@
 import { useIsLoggedIn } from "@/features/auth/store/useAuthStore";
-import { getFridgeApi } from "@/features/home/apis/fridge";
+import { getFridgeApi } from "@/features/fridge/apis/fridge";
 import { useApiQuery } from "@/shared/apis/builder/ApiBuilder";
 import { QUERY_KEYS } from "@/shared/constants/queryKeys";
