+++ src/features/home/apis/food.types.ts
@@ -1,20 +1,9 @@
 import { FoodItem, StorageType } from "@/shared/types/food";
 
-type FoodStatusData = {
-  black: FoodItem[];
-  red: FoodItem[];
-  yellow: FoodItem[];
-  green: FoodItem[];
-  blackCount: number;
-  redCount: number;
-  yellowCount: number;
-  greenCount: number;
-};
-
-interface FoodStatusResponse {
-  result: string;
-  data: FoodStatusData;
-}
+export type {
+  FoodStatusData,
+  FoodStatusResponse,
+} from "@/features/food/apis/foodStatus.types";
 
 type FoodCursorSortBy = "EXPIRATION" | "CREATED" | "NAME";
 
@@ -39,10 +28,4 @@ interface FridgeFoodsResponseRaw {
   hasMore?: boolean;
 }
 
-export type {
-  FoodCursorSortBy,
-  FoodStatusData,
-  FoodStatusResponse,
-  FridgeFoodsCursorRequest,
-  FridgeFoodsResponseRaw,
-};
+export type { FoodCursorSortBy, FridgeFoodsCursorRequest, FridgeFoodsResponseRaw };
