+++ src/features/home/hooks/queries/useAllFoodStatusQuery.ts
@@ -1,8 +1,8 @@
 import { useIsLoggedIn } from "@/features/auth/store/useAuthStore";
 import { useApiQuery } from "@/shared/apis/builder/ApiBuilder";
 import { QUERY_KEYS } from "@/shared/constants/queryKeys";
-import { getFoodStatusApi } from "../../apis/food";
-import { FoodStatusResponse } from "../../apis/food.types";
+import { getFoodStatusApi } from "@/features/food/apis/foodStatus";
+import type { FoodStatusResponse } from "@/features/food/apis/foodStatus.types";
 import { normalizeFoodItem } from "../../utils/normalizeFoodItem";
 
 const useAllFoodStatusQuery = (enabled = true) => {
